Update and rename CountriesTableSeeder.php to CountriesSeeder.php
To use this seeder in production, I've conformed it on a few fronts. First, the migration defines an `->increments('id')` column, so we do not need to supply this in the data arrays. Next, `name` was merely an uppercase clone of `nicename` with a few additional formatting rules (eg. no " / " in the string value). It has been removed and `nicename` has been renamed to the straightforward `name`, as a user can simply call `Str::upper()` on the `name` value instead of duplicating persisted data. Also, any database relationships tying back to any of these records should be linking via `id` or via the `iso` or `iso3` value, not via an uppercase string value (i.e. the no-longer-existing previous `name` value).